refactor(#67): splitter ShoppingPage et PlanningPage en sous-composants
ShoppingPage: 1276 → 411 lignes (LabelDropdown, FormLabelSelect, MealieItemRow,
HabituelItemRow, GroupedItems, GroupedHabituels, DefaultCatalogModal, buildPrintHtml)

PlanningPage: 1386 → 847 lignes (MobileMealSection, MealCell, planningUtils)

Closes #67
